refactor(local-ci): extract cloud provider integration into cloud.py (R2-1, #2645)
local_ci.py was 10,515 lines. Extract the cloud layer — GitHub CLI plumbing
(gh_*), Namespace plumbing (nsc_*), cloud-record / billing / provider-metadata
helpers, provider compare/recommend, and the cmd_cloud_* subcommands — into a
new tools/local-ci/cloud.py (~2,000 lines). local_ci.py: 10,515 -> 8,710.

The full cloud layer moves together (per the Codex consult: partial extraction
would create a local_ci <-> cloud import cycle). parse_iso_datetime (cloud-only)
moves too. cloud.py imports only leaf modules (state_paths/io_utils/git_helpers/
normalize/github_workflows) + stdlib; the one shared config helper
load_optional_config is imported lazily (_load_optional_config) to avoid the
cycle. local_ci.py re-exports the 82 public cloud symbols ("re-exported for
in-file consumers") so ship/check/list/desktop commands + main() dispatch are
unchanged.

Tests: the cloud tests monkeypatch helpers via direct attribute assignment
(self.mod.<fn> = ...); since cmd_cloud_* now resolve those in the cloud
namespace, the patches are redirected to a new self.cloud module reference
(incl. the re-imported leaf helper now_iso in the cloud-run tests). Full
tools/local-ci suite: 224/224 pass; `python local_ci.py --help` byte-stable.

Skill-Update: skip skill=ci reason="mechanical module extraction from local_ci.py; no CI workflow, command, or behavior change — the python local_ci.py entrypoint and all subcommands are byte-stable"
